fix num_stages and move function to utils
